% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{residual_plot}
\alias{residual_plot}
\title{Residual Plot for Simple Linear Models}
\usage{
residual_plot(df, y, b1)
}
\arguments{
\item{df}{Data frame being used, hence df}

\item{y}{Numeric vector from df, which is the response variable, or "y" variable, need to be in df$column_name}

\item{b1}{Numeric Vector from df, which is the explanatory variable. In simple linear regression, the slope associated with the first x variable would be b1, need to be in ""}
}
\value{
A graph
}
\description{
This model creates a residual plot for a linear model with 1 explanatory variable. This can be used to test if the model meets the linearity and equal variance assumptions
}
\examples{
residual_plot(mtcars, mtcars$mpg, mtcars$wt) #it plots

}
